edit admin thống kê, chưa lưu dc hóa đơn xuống csdl nhưng sao trên quản lý nó vẫn hiên, nó lưu ở đâu không bt
