+++ QDMarketPlace.Data.EF/appsettings.json
@@ -1,6 +1,6 @@
 ﻿{
   "ConnectionStrings": {
-    "DefaultConnection": "Server=tcp:qdmarket.database.windows.net,1433;Initial Catalog=QDMarketDB_6_7;Persist Security Info=False;User ID=quocitspkt;Password=Dauchodau1;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+    "DefaultConnection": "Server=DESKTOP-PQJNLH9\\SQLEXPRESS;Database=QDMarketDB;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "PageSize": 12,
   "Logging": {
