+++ QDMarketPlace/Areas/Admin/Views/Login/Index.cshtml
@@ -39,10 +39,10 @@
                         @Html.AntiForgeryToken()
                         <h1>Đăng nhập</h1>
                         <div>
-                            <input type="text" class="form-control" name="userName" placeholder="User name" id="txtUserName" required="" />
+                            <input type="text" class="form-control" name="userName" placeholder="Tên người dùng" id="txtUserName" required="" />
                         </div>
                         <div>
-                            <input type="password" class="form-control" name="password" id="txtPassword" placeholder="Password" required="" />
+                            <input type="password" class="form-control" name="password" id="txtPassword" placeholder="Mật khẩu" required="" />
                         </div>
                         <div>
                             <a class="btn btn-default submit" href="#" id="btnLogin">Đăng nhập</a>
