+++ QDMarketPlace/Controllers/CartController.cs
@@ -414,22 +414,22 @@ namespace QDMarketPlace.Controllers
                         item.Product = product;
                         item.Size = _billService.GetSize(size);
                         item.Color = _billService.GetColor(color);
-                        if (quantity > 6)
-                        {
-                            item.Quantity = 6;
-                        }
-                        else
-                        {
-                            item.Quantity = quantity;
-                        }
-                        //if (quantity > _productService.GetAmount(productId))
+                        //if (quantity > 6)
                         //{
-                        //    item.Quantity = _productService.GetAmount(productId);
+                        //    item.Quantity = 6;
                         //}
                         //else
                         //{
                         //    item.Quantity = quantity;
                         //}
+                        if (quantity > _productService.GetAmount(productId))
+                        {
+                            item.Quantity = _productService.GetAmount(productId);
+                        }
+                        else
+                        {
+                            item.Quantity = quantity;
+                        }
                         item.Price = product.PromotionPrice ?? product.Price;
                         hasChanged = true;
                     }
