+++ QDMarketPlace/Views/Shared/_Header.cshtml
@@ -63,11 +63,11 @@
                             <form method="GET" action="/search.html" class="">
                                 <div class="clearfix category-box relative">
                                     <select name="cate_search">
-                                        <option>All Category</option>
+                                        <option>Loại sản phẩm</option>
                                     </select>
                                 </div>
                                 <input type="text" name="keyword" placeholder="Nhập từ khóa cần tìm . . .">
-                                <button type="submit" class="animate-default button-hover-red">Tìm Kiếm</button>
+                                <button type="submit" class="animate-default button-hover-red" value="Tìm kiếm">Tìm Kiếm</button>
                             </form>
                         </div>
                         <div class="clearfix icon-search-mobile absolute">
