+++ QDMarketPlace/appsettings.json
@@ -1,6 +1,6 @@
 ﻿{
   "ConnectionStrings": {
-    "DefaultConnection": "Server=DESKTOP-MUGMI18\\MSSQLSERVERQUOC;Database=QDMarketDB;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "DefaultConnection": "Server=DESKTOP-PQJNLH9\\SQLEXPRESS;Database=QDMarketDB;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "PageSize": 12,
   "Logging": {
